pyproject.toml
[tool.poetry]
name = "askatt-experiments"
version = "0.0.1"
description = ""
authors = ["Nair, Maitreyi (MSFT) <ms5253@att.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
azure-identity = "^1.14.1"
azure-storage-blob = "^12.18.3"
evaluate = "^0.4.1"
rouge-score = "^0.1.2"
absl-py = "^2.0.0"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"
sentence-transformers = "^2.2.2"
torch = "^2.0.1"
matplotlib = "^3.8.0"
loguru = "^0.7.2"
requests = "^2.31.0"
pydantic = "^1.10.13"
beautifulsoup4 = "^4.12.2"
plotly = "^5.18.0"
ipykernel = "^6.26.0"
langchain = "^0.0.341"
openai = "^1.3.7"


[[tool.poetry.source]]
name = "attrepo"
url = "https://pypi.it.att.com"
priority = "default"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"


[tool.poetry.group.ingest.dependencies]
snowflake-connector-python = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
